desktopmac/assets/ComfyUI/requirements.txt
# ComfyUI Lite 自定义包（从 GitHub Release 安装）
comfyui-lite-frontend @ https://github.com/josephxie1/ComfyUi-Lite/releases/download/v1.0.0/comfyui_lite_frontend-1.0.0-py3-none-any.whl
comfyui-workflow-templates-core @ https://github.com/josephxie1/ComfyUi-Lite/releases/download/v1.0.0/comfyui_workflow_templates_core-0.3.166-py3-none-any.whl
comfyui-workflow-templates-media-api @ https://github.com/josephxie1/ComfyUi-Lite/releases/download/v1.0.0/comfyui_workflow_templates_media_api-0.3.62-py3-none-any.whl
comfyui-embedded-docs==0.4.3

# Core runtime
numpy>=1.25.0
aiohttp>=3.11.8
yarl>=1.18.0
pyyaml
Pillow
tqdm
psutil
requests
simpleeval>=1.0.0

# Database
alembic
SQLAlchemy

# Comfy platform
comfy-kitchen>=0.2.7
comfy-aimdo>=0.2.7

# Optional
pydantic~=2.0
pydantic-settings~=2.0
